+++ blazor-application/.idea/.idea.BlazorBLE/.idea/.name
@@ -1,0 +1,1 @@
+BlazorBLE
